fix: solve the problem of circular import;
moved the functions of parsing delta to util.py.
