+++ pypots/data/__init__.py
@@ -9,16 +9,19 @@ from pypots.data.base import BaseDataset
 from pypots.data.dataset_for_brits import DatasetForBRITS
 from pypots.data.dataset_for_grud import DatasetForGRUD
 from pypots.data.dataset_for_mit import DatasetForMIT
+
 from pypots.data.generating import (
     generate_random_walk,
     generate_random_walk_for_classification,
 )
-from pypots.data.integration import (
+
+from pypots.data.utils import (
     masked_fill,
     mcar,
     pickle_load,
     pickle_dump,
 )
+
 from pypots.data.load_specific_datasets import (
     list_supported_datasets,
     load_specific_dataset,
